refactor: extract LogEntry component from erg-dashboard.jsx
Moves 310-line shared component to web/src/components/LogEntry.jsx.
Used across plan, overview, erg, strength, and log views.

Build and 68 tests pass.
